Install playwright browser dependencies
This feature depends on having the ghcr.io/devcontainers/features/node
feature installed in the devcontainer.

I've specified that it "installsAfter" the node feature so that npm is
available, allowing us to use it to install the playwright command.

I didn't use "dependsOn", as I didn't the use of this feature to cause
people who are using it to have to pull down two different versions of
that node feature (one with no options set, and one that might specify
an options object like this:

    {"version": "20.11.0"}

If that happens then devcontainer/cli will consider them to be separate
features. So we'll have to make sure that people appreciate that they
need to pull down Node themselves, which I'd expect most of them are
probably already doing anyway.

Note that we uninstall playwright after using it to install
dependencies, so that it doesn't interfere with the version of
playwright that is installed into the user's ./node_modules directory,
as part of their project's build.