Make browser cache directory writable
I tried doing this by running chown directly inside install.sh,
referencing the $_CONTAINER_USER variable. That doesn't work, as the
volume isn't mounted at build time. Obviously!

The approach taken was inpsired by the Rails bundler-cache feature [1],
and has the advantage that if you launch a container as a different
user, ownership in the volume will get updated to match.

Note that the postCreateCommand used checks whether you are already root
before running chown, and will run `true` if you are root. This is to
ensure that it exits cleanly when executed by the tests, that execute
the postCreateCommand as root.

[1] https://github.com/rails/devcontainer/commit/073e4ba2002b7959c2eb00968ab40ec8946392d4