Install Playwright as container's remote user
We were previously installing them as the "container user", which
equates to whichever user was specified in the Dockerfile, defaulting to
root.

Dev containers are often setup to run processes within the container as
a separate user, set via the remoteUser property in devcontainer.json.
It's often set to "vscode".

It's typical to run npm and npx as a non-root user, hence this change.
Note that `npx playwright install-deps` runs sudo automatically.